+++ CHANGELOG.md
@@ -2,4 +2,4 @@ V0.1
 
 - Fix inconsistencies in button & input styling, refactoring of styles
 - Added feature of logo image upload
-- Fix missing inputs for background color
+- Fix missing inputs for background color and border-radius